+++ src/Services/Customer.API/Dockerfile
@@ -1,0 +1,25 @@
+FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
+WORKDIR /app
+EXPOSE 80
+
+FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
+WORKDIR /src
+
+COPY ["Services/Customer.API/Customer.API.csproj", "Services/Customer.API/"]
+COPY ["BuildingBlocks/Common.Logging/Common.Logging.csproj", "BuildingBlocks/Common.Logging/"]
+COPY ["BuildingBlocks/Contracts/Contracts.csproj", "BuildingBlocks/Contracts/"]
+COPY ["BuildingBlocks/Infrastructure/Infrastructure.csproj", "BuildingBlocks/Infrastructure/"]
+COPY ["BuildingBlocks/Shared/Shared.csproj", "BuildingBlocks/Shared/"]
+
+RUN dotnet restore "Services/Customer.API/Customer.API.csproj"
+COPY . .
+WORKDIR "/src/Services/Customer.API"
+RUN dotnet build "Customer.API.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "Customer.API.csproj" -c Release -o /app/publish
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "Customer.API.dll"]
