+++ src/Services/Customer.API/Extensions/ApplicationExtension.cs
@@ -6,7 +6,11 @@
         {
 
             app.UseSwagger();
-            app.UseSwaggerUI();
+            app.UseSwaggerUI(c =>
+            {
+                c.SwaggerEndpoint("/swagger/v1/swagger.json",
+                    "Swagger Customer Minial API v1");
+            });
 
             app.UseRouting();
             // app.UseHttpsRedirection(); //for production only
