+++ src/Services/Customer.API/Repositories/CustomerRepository.cs
@@ -6,9 +6,9 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Customer.API.Repositories
 {
-        public class CustomerRepository : RepositoryBaseAsync<Entities.Customer, int, CustomerContext>, ICustomerRepository
+        public class CustomerRepository : RepositoryQueryBase<Entities.Customer, int, CustomerContext>, ICustomerRepository
         {
-        public CustomerRepository(CustomerContext dbContext, IUnitOfWork<CustomerContext> unitOfWork) : base(dbContext, unitOfWork)
+        public CustomerRepository(CustomerContext dbContext) : base(dbContext)
         {
         }
