+++ src/docker-compose.override.yml
@@ -105,7 +105,7 @@ services:
       - elasticsearch
     ports:
       - "5601:5601"
-      
+            
       
   #microservices api sections
   product.api:
@@ -119,15 +119,15 @@ services:
     ports:
       - "6002:80"
     restart: always
-    
-  #customer.api:
-  #  container_name: customer.api
-  #  environment:
-  #    - ASPNETCORE_ENVIRONMENT=Development
-  #    - ASPNETCORE_URLS=http://+:80
-  #    - "ConnectionStrings:DefaultConnectionString=Server=customerdb;Port=5432;Database=CustomerDb;User Id=admin;Password=admin1234"
-  #  depends_on: 
-  #    - customerdb
-  #  ports:
-  #    - "6003:80"
-  #  restart: always
+
+  customer.api:
+    container_name: customer.api
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Development
+      - ASPNETCORE_URLS=http://+:80
+      - "ConnectionStrings:DefaultConnectionString=Server=customerdb;Port=5432;Database=CustomerDb;User Id=admin;Password=admin1234"
+    depends_on: 
+      - customerdb
+    ports:
+      - "6003:80"
+    restart: always
