+++ src/docker-compose.yml
@@ -39,7 +39,12 @@ services:
     build:
       context: .
       dockerfile: Services/Product.API/Dockerfile
-    
+
+  customer.api:
+    image: ${DOCKER_REGISTRY-}customer-api:${PLATFORM:-linux}-${TAG:-latest}
+    build:
+      context: .
+      dockerfile: Services/Customer.API/Dockerfile
     
 #network section
 networks:
